Updating PR based on comments

+++ controllers/order.js
@@ -22,7 +22,7 @@ exports.getAnOrder = asyncHandler(async (req , res , next)=>{
     const orderId = req.params.id;
     const foundOrder = await Order.findById(orderId);
     if(!foundOrder){
-        res.status(500).json({error : "something went wrong"});
+        res.status(404).json({error : "Resource not found"});
     }else{
         res.status(200).json({order : foundOrder})
     }
@@ -33,11 +33,16 @@ exports.getAnOrder = asyncHandler(async (req , res , next)=>{
 // @access Private
 exports.updateOrder = asyncHandler(async (req , res , next)=>{
     const orderId = req.params.id;
+   
+    //check if order exists
+    const order = await Order.findById(orderId);
+    if(!order) return res.status(404).json({error : "Order not found"})
+
     updatedOrder = await Order.findByIdAndUpdate(orderId , req.body);
     if(!updatedOrder){
-        res.status(500).json({error : "Something went wrong"})
+        res.status(500).json({error : "Something went wrong"});
     }else{
-        res.status(200).json({order : updatedOrder})
+        res.status(200).json({order : updatedOrder});
     }
 })
 
@@ -46,9 +51,11 @@ exports.updateOrder = asyncHandler(async (req , res , next)=>{
 // @access Private
 exports.deleteOrder = asyncHandler(async (req , res , next)=>{
     const orderId = req.params.id;
-    console.log(orderId)
+
+    const order = await Order.findById(orderId);
+    if(!order) return res.status(404).json({error : "Order not found"})
+    
     const deletedOrder = await Order.findByIdAndDelete(orderId);
-    console.log(deletedOrder)
     if(!deletedOrder){
         res.status(500).json({error : "Something went wrong"})
     }else{
@@ -74,26 +81,18 @@ exports.getUserOrders = asyncHandler(async(req , res , next)=>{
 // @access Private (Admin)
 exports.getAllOrders = asyncHandler(async(req , res , next)=>{
     const orders = await Order.find()
-    if(!orders){
-        res.status(500).json({error : "Something went wrong"})
-    }else{
-        res.status(200).json({orders : orders})
-    }
+    res.status(200).json({orders : orders});
 })
 
 exports.getMonthlyIncome = asyncHandler(async(req , res , next)=>{
     const date = new Date();
     const lastMonth = new Date(date.setMonth(date.getMonth() - 1));  
     const previousMonth = new Date(new Date().setMonth(lastMonth.getMonth() - 1));
-    try{
-        const income = await Order.aggregate([
-            {$match : {createdAt : {$gte : previousMonth}}},
-            {$project : {month : {$month : "$createdAt"} , sales : "$amount"}},
-            {$group : {_id : "$month", total : {$sum : "$sales"}}}
-        ])
-        res.status(200).json(income)
-    }catch(err){
-        res.status(500).json(err)
-    } 
+    const income = await Order.aggregate([
+        {$match : {createdAt : {$gte : previousMonth}}},
+        {$project : {month : {$month : "$createdAt"} , sales : "$amount"}},
+        {$group : {_id : "$month", total : {$sum : "$sales"}}}
+    ])
+    res.status(200).json(income);
 })
